Reorganised files + started on data layer logic
